refactor(#445): split the Campaign RPC surface into feature modules over narrow store slices
The 39-method campaignStore god-interface is gone. CampaignServer now
composes eight feature modules - campaign management, archive lifecycle,
Agent roster, Player Characters, KG Nodes, KG Edges, Knowledge Proposals,
Tool Grants - each declaring the 3-6-method store slice it actually
consumes (consumer-defined, satisfied structurally by *storage.Store) and
promoting its RPC methods onto the unchanged CampaignServiceHandler. A
shared *activeCampaignSource carries the live-session closure plus the
3-method resolver, so SetSessions still wires every surface at once. No
proto/wire change; main.go wiring is untouched (NewCampaignServer keeps
its call shape over *storage.Store); NewCampaignServerWith(CampaignStores)
composes per-feature slices for tests.

Tests: the two 39-method fakes (fakeReader's 36 dead stubs,
fakeCampaignStore's ~70 config fields) are replaced by per-feature fakes
in fakes_test.go, each embedding the 3-method fakeActive; all 170 unit
tests are preserved verbatim, only construction plumbing changed. The
Postgres-bound suites shrink 15 -> 10: each deleted test has a keyless
unit twin (GetActiveCampaign mapping, live-first SetActiveCampaign,
cross-campaign grant/edge reads) or storage-level SQL coverage, and the
#268 language-edit/voice DB guard moves to internal/storage
(TestUpdateCampaignLanguageLeavesAgentVoiceUntouched) where the SQL is
the subject.

The largest store interface in internal/rpc is now 6 methods, and the
package doc's "narrow reader interfaces" claim is finally true for this
server.

Closes #445
