+++ internal/rpc/active_campaign.go
@@ -65,3 +65,36 @@ func resolveActiveCampaign(ctx context.Context, live func() (uuid.UUID, bool), s
 	}
 	return store.GetActiveCampaign(ctx)
 }
+
+// activeCampaignSource is the ONE Active-Campaign resolution every CampaignServer
+// feature module shares (#445): the live Voice Session closure SetSessions wires
+// plus the 3-method resolver slice. The modules hold a pointer to the same
+// instance, so the header, CRUD, KG, proposal, and grant surfaces always resolve
+// the same campaign — and SetSessions wiring the closure once at boot reaches all
+// of them.
+type activeCampaignSource struct {
+	// live reports the live Voice Session's campaign id, if any. Nil until
+	// SetSessions wires it (keyless deployments and most unit tests leave it nil,
+	// which skips the live-first step). Set once at boot before serving, so no
+	// lock is needed.
+	live  func() (uuid.UUID, bool)
+	store activeCampaignResolver
+}
+
+// resolve resolves the operator's Active Campaign via the one shared
+// resolveActiveCampaign policy (live Voice Session → durable /glyphoxa use
+// selection → most-recent fallback, #222).
+func (a *activeCampaignSource) resolve(ctx context.Context) (storage.Campaign, error) {
+	return resolveActiveCampaign(ctx, a.live, a.store)
+}
+
+// liveID reports the LIVE Voice Session's campaign id, nil-safe for an unwired
+// source. The archive/delete live-guard reads it directly — the campaign that is
+// currently voicing must not be archived or deleted, whether or not it is the
+// resolved Active Campaign (#265).
+func (a *activeCampaignSource) liveID() (uuid.UUID, bool) {
+	if a.live == nil {
+		return uuid.Nil, false
+	}
+	return a.live()
+}
