+++ internal/rpc/character_test.go
@@ -8,69 +8,26 @@ import (
 	"github.com/google/uuid"
 
 	managementv1 "github.com/MrWong99/Glyphoxa/gen/glyphoxa/management/v1"
+	"github.com/MrWong99/Glyphoxa/gen/glyphoxa/management/v1/managementv1connect"
+	"github.com/MrWong99/Glyphoxa/internal/rpc"
 	"github.com/MrWong99/Glyphoxa/internal/storage"
 )
 
-// --- fakeCampaignStore Player Character methods (#276) ---
-
-func (f *fakeCampaignStore) ListCharacters(_ context.Context, campaignID uuid.UUID) ([]storage.Character, error) {
-	f.charsListCampaign = campaignID
-	return f.characters, f.charListErr
-}
-
-func (f *fakeCampaignStore) CreateCharacter(_ context.Context, c storage.NewCharacter) (uuid.UUID, error) {
-	if f.charCreateErr != nil {
-		return uuid.Nil, f.charCreateErr
-	}
-	f.charsCreated = append(f.charsCreated, c)
-	id := uuid.New()
-	f.characters = append(f.characters, storage.Character{
-		ID:            id,
-		CampaignID:    c.CampaignID,
-		Name:          c.Name,
-		Aliases:       c.Aliases,
-		DiscordUserID: c.DiscordUserID,
-	})
-	return id, nil
-}
-
-func (f *fakeCampaignStore) UpdateCharacter(_ context.Context, u storage.CharacterUpdate) (storage.Character, error) {
-	f.charUpdateCampaign = u.CampaignID
-	if f.charUpdateErr != nil {
-		return storage.Character{}, f.charUpdateErr
-	}
-	for i := range f.characters {
-		if f.characters[i].ID == u.ID {
-			f.characters[i].Name = u.Name
-			f.characters[i].Aliases = u.Aliases
-			f.characters[i].DiscordUserID = u.DiscordUserID
-			return f.characters[i], nil
-		}
-	}
-	return storage.Character{}, storage.ErrNotFound
-}
-
-func (f *fakeCampaignStore) DeleteCharacter(_ context.Context, campaignID, id uuid.UUID) error {
-	f.charDeleteCampaign = campaignID
-	if f.charDeleteErr != nil {
-		return f.charDeleteErr
-	}
-	for i := range f.characters {
-		if f.characters[i].ID == id {
-			f.characters = append(f.characters[:i], f.characters[i+1:]...)
-			return nil
-		}
-	}
-	return storage.ErrNotFound
+// characterClient composes a CampaignServer from the Player Character fake —
+// filling only the Active + Characters slices (#445) — and returns the
+// Connect-JSON client.
+func characterClient(t *testing.T, store *fakeCharacterStore) managementv1connect.CampaignServiceClient {
+	t.Helper()
+	return campaignClient(t, rpc.CampaignStores{Active: store, Characters: store})
 }
 
 // --- CreateCharacter ---
 
 func TestCreateCharacter_MapsAndPersists(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campaign = storage.Campaign{ID: uuid.New(), Name: "Lost Mine"}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	resp, err := client.CreateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.CreateCharacterRequest{
@@ -108,9 +65,9 @@ func TestCreateCharacter_MapsAndPersists(t *testing.T) {
 
 func TestCreateCharacter_EmptyNameIsInvalidArgument(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campaign = storage.Campaign{ID: uuid.New()}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.CreateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.CreateCharacterRequest{
@@ -124,9 +81,9 @@ func TestCreateCharacter_EmptyNameIsInvalidArgument(t *testing.T) {
 func TestCreateCharacter_NonSnowflakeIsInvalidArgument(t *testing.T) {
 	t.Parallel()
 	for _, bad := range []string{"", "not-a-number", "123abc", "-5", "12 34", "12.3"} {
-		store := newFakeStore()
+		store := newFakeCharacterStore()
 		store.campaign = storage.Campaign{ID: uuid.New()}
-		client := crudClient(t, store)
+		client := characterClient(t, store)
 
 		_, err := client.CreateCharacter(context.Background(),
 			connect.NewRequest(&managementv1.CreateCharacterRequest{
@@ -143,10 +100,10 @@ func TestCreateCharacter_NonSnowflakeIsInvalidArgument(t *testing.T) {
 
 func TestCreateCharacter_DuplicateIsAlreadyExists(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campaign = storage.Campaign{ID: uuid.New()}
 	store.charCreateErr = storage.ErrConflict
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.CreateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.CreateCharacterRequest{
@@ -159,9 +116,9 @@ func TestCreateCharacter_DuplicateIsAlreadyExists(t *testing.T) {
 
 func TestCreateCharacter_NoCampaignIsNotFound(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campErr = storage.ErrNotFound
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.CreateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.CreateCharacterRequest{
@@ -176,13 +133,13 @@ func TestCreateCharacter_NoCampaignIsNotFound(t *testing.T) {
 
 func TestListCharacters_ScopedToActiveCampaign(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	activeID := uuid.New()
 	store.campaign = storage.Campaign{ID: activeID}
 	store.characters = []storage.Character{
 		{ID: uuid.New(), CampaignID: activeID, Name: "Aravel", DiscordUserID: "1"},
 	}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	resp, err := client.ListCharacters(context.Background(),
 		connect.NewRequest(&managementv1.ListCharactersRequest{}))
@@ -199,9 +156,9 @@ func TestListCharacters_ScopedToActiveCampaign(t *testing.T) {
 
 func TestListCharacters_NoCampaignIsNotFound(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campErr = storage.ErrNotFound
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.ListCharacters(context.Background(),
 		connect.NewRequest(&managementv1.ListCharactersRequest{}))
@@ -214,14 +171,14 @@ func TestListCharacters_NoCampaignIsNotFound(t *testing.T) {
 
 func TestUpdateCharacter_Rebinds(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	activeID := uuid.New()
 	store.campaign = storage.Campaign{ID: activeID}
 	id := uuid.New()
 	store.characters = []storage.Character{
 		{ID: id, CampaignID: activeID, Name: "Old", DiscordUserID: "111111111111111111"},
 	}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	resp, err := client.UpdateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.UpdateCharacterRequest{
@@ -241,9 +198,9 @@ func TestUpdateCharacter_Rebinds(t *testing.T) {
 
 func TestUpdateCharacter_InvalidIdIsInvalidArgument(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campaign = storage.Campaign{ID: uuid.New()}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.UpdateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.UpdateCharacterRequest{
@@ -256,9 +213,9 @@ func TestUpdateCharacter_InvalidIdIsInvalidArgument(t *testing.T) {
 
 func TestUpdateCharacter_UnknownIdIsNotFound(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campaign = storage.Campaign{ID: uuid.New()}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.UpdateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.UpdateCharacterRequest{
@@ -271,10 +228,10 @@ func TestUpdateCharacter_UnknownIdIsNotFound(t *testing.T) {
 
 func TestUpdateCharacter_ConflictIsAlreadyExists(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campaign = storage.Campaign{ID: uuid.New()}
 	store.charUpdateErr = storage.ErrConflict
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.UpdateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.UpdateCharacterRequest{
@@ -287,9 +244,9 @@ func TestUpdateCharacter_ConflictIsAlreadyExists(t *testing.T) {
 
 func TestUpdateCharacter_NonSnowflakeIsInvalidArgument(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campaign = storage.Campaign{ID: uuid.New()}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.UpdateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.UpdateCharacterRequest{
@@ -305,9 +262,9 @@ func TestUpdateCharacter_NonSnowflakeIsInvalidArgument(t *testing.T) {
 // CodeNotFound rather than mutating by bare id.
 func TestUpdateCharacter_NoActiveCampaignIsNotFound(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campErr = storage.ErrNotFound
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.UpdateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.UpdateCharacterRequest{
@@ -322,10 +279,10 @@ func TestUpdateCharacter_NoActiveCampaignIsNotFound(t *testing.T) {
 
 func TestDeleteCharacter_Removes(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	id := uuid.New()
 	store.characters = []storage.Character{{ID: id, Name: "Doomed", DiscordUserID: "1"}}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	if _, err := client.DeleteCharacter(context.Background(),
 		connect.NewRequest(&managementv1.DeleteCharacterRequest{Id: id.String()})); err != nil {
@@ -338,8 +295,8 @@ func TestDeleteCharacter_Removes(t *testing.T) {
 
 func TestDeleteCharacter_InvalidIdIsInvalidArgument(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
-	client := crudClient(t, store)
+	store := newFakeCharacterStore()
+	client := characterClient(t, store)
 
 	_, err := client.DeleteCharacter(context.Background(),
 		connect.NewRequest(&managementv1.DeleteCharacterRequest{Id: "bad"}))
@@ -350,8 +307,8 @@ func TestDeleteCharacter_InvalidIdIsInvalidArgument(t *testing.T) {
 
 func TestDeleteCharacter_UnknownIdIsNotFound(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
-	client := crudClient(t, store)
+	store := newFakeCharacterStore()
+	client := characterClient(t, store)
 
 	_, err := client.DeleteCharacter(context.Background(),
 		connect.NewRequest(&managementv1.DeleteCharacterRequest{Id: uuid.NewString()}))
@@ -365,14 +322,14 @@ func TestDeleteCharacter_UnknownIdIsNotFound(t *testing.T) {
 // campaign_id) and a cross-campaign write is refused server-side.
 func TestUpdateCharacter_ScopesToActiveCampaign(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	activeID := uuid.New()
 	store.campaign = storage.Campaign{ID: activeID}
 	id := uuid.New()
 	store.characters = []storage.Character{
 		{ID: id, CampaignID: activeID, Name: "Old", DiscordUserID: "111111111111111111"},
 	}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	if _, err := client.UpdateCharacter(context.Background(),
 		connect.NewRequest(&managementv1.UpdateCharacterRequest{
@@ -389,12 +346,12 @@ func TestUpdateCharacter_ScopesToActiveCampaign(t *testing.T) {
 // resolved active campaign, so another campaign's Character is never removable.
 func TestDeleteCharacter_ScopesToActiveCampaign(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	activeID := uuid.New()
 	store.campaign = storage.Campaign{ID: activeID}
 	id := uuid.New()
 	store.characters = []storage.Character{{ID: id, CampaignID: activeID, Name: "Doomed", DiscordUserID: "1"}}
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	if _, err := client.DeleteCharacter(context.Background(),
 		connect.NewRequest(&managementv1.DeleteCharacterRequest{Id: id.String()})); err != nil {
@@ -409,9 +366,9 @@ func TestDeleteCharacter_ScopesToActiveCampaign(t *testing.T) {
 // campaign the scoped delete cannot resolve an owner and returns CodeNotFound.
 func TestDeleteCharacter_NoActiveCampaignIsNotFound(t *testing.T) {
 	t.Parallel()
-	store := newFakeStore()
+	store := newFakeCharacterStore()
 	store.campErr = storage.ErrNotFound
-	client := crudClient(t, store)
+	client := characterClient(t, store)
 
 	_, err := client.DeleteCharacter(context.Background(),
 		connect.NewRequest(&managementv1.DeleteCharacterRequest{Id: uuid.NewString()}))
