+++ internal/rpc/voice_members.go
@@ -22,7 +22,7 @@ type voiceMemberLister func(ctx context.Context) ([]presence.Member, error)
 // needed. Left nil on a keyless / bot-offline deployment, in which case
 // ListDiscordVoiceMembers returns an empty list.
 func (s *CampaignServer) SetMemberLister(l voiceMemberLister) {
-	s.memberLister = l
+	s.characterRoster.memberLister = l
 }
 
 // ListDiscordVoiceMembers returns the Discord Users currently in the operator's
@@ -32,7 +32,7 @@ func (s *CampaignServer) SetMemberLister(l voiceMemberLister) {
 // and never an error, so the picker degrades to free-text snowflake entry
 // (ADR-0003). No privileged intent is used — the members come from the
 // voice-state cache the Bot already receives.
-func (s *CampaignServer) ListDiscordVoiceMembers(
+func (s *characterRoster) ListDiscordVoiceMembers(
 	ctx context.Context,
 	_ *connect.Request[managementv1.ListDiscordVoiceMembersRequest],
 ) (*connect.Response[managementv1.ListDiscordVoiceMembersResponse], error) {
